Add files via upload
We will perform multiple sequence alignment and differential expressional analysis and use a heatmap and MSA pretty print to visualize how similar the human CMAH gene expression is to species like chimpanzees and mice. Multiple sequence alignment is performed to compare the regions in which CMAH is present or mutated and the data was gathered through the NCBI nucleotide. Differential expression analysis is visually represented through a heatmap to understand whether CMAH is expressed across the different species. We are using this analysis to approach the scientific hypothesis: If there is a deletion of the 92-bp in the CMAH gene, then the CMAH gene in humans would not be expressed indicating no production of sialic acids in humans and less similarity to CMAH genes present in mice or chimpanzees. 

README files:
Project 2 code draft rmd/html file: very rough draft, still trying to get code to work of my project 2
Mice_CMAH fasta: sequence coding CMAH gene in mice
Humans_CMAH fasta: sequence coding CMAH gene in humans (although nonfunctional due to mutation)